+++ packages/wagmi/src/wagmi-provider/__tests__/balance.test.tsx
@@ -11,7 +11,7 @@ const mockConnector = {
   name: 'MetaMask',
 };
 
-vi.mock('@wagmi/core', () => ({
+vi.mock('wagmi/actions', () => ({
   getAccount: () => ({}),
   disconnect: () => {},
 }));
@@ -54,6 +54,7 @@ vi.mock('wagmi', () => {
       };
     },
     useEnsName: () => ({}),
+    useEnsAvatar: () => ({}),
   };
 });