+++ packages/wagmi/src/wagmi-provider/__tests__/switch-chain.test.tsx
@@ -16,7 +16,7 @@ const mockConnector = {
   name: 'MetaMask',
 };
 
-vi.mock('@wagmi/core', () => ({
+vi.mock('wagmi/actions', () => ({
   getAccount: () => ({}),
   disconnect: () => {},
 }));
@@ -70,6 +70,7 @@ vi.mock('wagmi', () => {
       };
     },
     useEnsName: () => ({}),
+    useEnsAvatar: () => ({}),
   };
 });